apps/racing-coach-server/pyproject.toml
[project]
name = "racing-coach-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Sawyer", email = "sawyerczupka@gmail.com" }]
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.14.0",
    "argon2-cffi>=23.1.0",
    "asyncpg>=0.30.0",
    "email-validator>=2.1.0",
    "fastapi>=0.115.12",
    "fastapi-cli>=0.0.7",
    "pydantic-settings>=2.9.1",
    "racing-coach-core",
    "sqlalchemy[asyncio]>=2.0.40",
    "uvicorn>=0.34.0",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.11",
]

[tool.uv.sources]
racing-coach-core = { path = "../../libs/racing-coach-core", editable = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
test = [
    "factory-boy>=3.3.3",
    "psycopg2-binary>=2.9.0",
    "pytest>=9.0.0",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-factoryboy>=2.8.1",
    "pytest-mock>=3.15.1",
    "pytest-randomly>=4.0.1",
    "testcontainers[postgres]>=4.13.3",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
markers = [
    "unit: Unit tests that mock external dependencies",
    "integration: Integration tests that test components together",
    "slow: Tests that take a long time to run",
]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=racing_coach_server",
    "--cov-report=term-missing",
    "--cov-report=html",
]
